# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIDB — API POS",
        "description": "API RESTful para el punto de venta (POS) del Sistema Integrado de Distribución y Bodega.<br><br>**Cómo autenticarse:**<br>1. Ejecuta **POST /login** con tu email y password.<br>2. Copia el `token` de la respuesta.<br>3. Haz clic en el botón **Authorize 🔒** (arriba a la derecha).<br>4. Escribe: `Bearer {tu-token}` y confirma.",
        "contact": {
            "name": "Soporte SIDB",
            "email": "soporte@sidb.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/SIDB/public/api",
            "description": "Laragon local"
        }
    ],
    "paths": {
...
        "/reintegros/candidatos": {
            "get": {
                "tags": [
                    "Reintegros"
                ],
                "summary": "Ventas candidatas a reintegro por cuotas vencidas",
                "description": "Devuelve ventas con cuotas vencidas hace N días (default 30), con los productos de la venta y datos del cliente. Filtra por sucursal del usuario autenticado.",
                "operationId": "90d3820952d2c97514e465acaf25f940",
                "parameters": [
                    {
                        "name": "dias_vencidos",
                        "in": "query",
                        "description": "Días mínimos de vencimiento para considerar candidato",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de candidatos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "venta_id": {
                                                        "type": "integer"
                                                    },
                                                    "numero_venta": {
                                                        "type": "string"
                                                    },
                                                    "fecha_venta": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "saldo_pendiente": {
                                                        "type": "number"
                                                    },
                                                    "cuotas_vencidas": {
                                                        "type": "integer"
                                                    },
                                                    "monto_adeudado_vencido": {
                                                        "type": "number"
                                                    },
                                                    "ya_en_reintegro": {
                                                        "type": "boolean"
                                                    },
                                                    "cliente": {
                                                        "type": "object"
                                                    },
                                                    "productos": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reintegros/vendedores": {
            "get": {
                "tags": [
                    "Reintegros"
                ],
                "summary": "Listar vendedores disponibles para asignar reintegro",
                "operationId": "381eee24b119752de52a037e0f438e0e",
                "responses": {
                    "200": {
                        "description": "Lista de vendedores activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "vendedores": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "codigo": {
                                                        "type": "string"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "telefono": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/reintegros/{id}/estado": {
            "patch": {
                "tags": [
                    "Reintegros"
                ],
                "summary": "Actualizar estado de un reintegro",
                "description": "Cambia el estado luego de la visita al cliente. Al marcar recuperado o no_recuperado se registra la fecha de recuperación automáticamente.",
                "operationId": "2afcc97861932b57e42135caf0aa976d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "recuperado",
                                        "enum": [
                                            "en_proceso",
                                            "recuperado",
                                            "no_recuperado"
                                        ]
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Se recogieron todos los productos en buen estado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "reintegro": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "estado": {
                                                    "type": "string"
                                                },
                                                "fecha_recuperacion": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Reintegro no encontrado"
                    },
                    "422": {
                        "description": "Estado inválido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Pega aquí el token obtenido en POST /login (sin la palabra Bearer)",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}